pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fedquery"
version = "0.1.0"
description = "FOMC Agentic RAG Research Assistant with citation-grounded answers"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "chromadb>=0.4",
    "faiss-cpu>=1.7",
    "sentence-transformers>=3.0",
    "mcp>=1.0",
    "typer>=0.12",
    "rich>=13.0",
    "beautifulsoup4>=4.12",
    "requests>=2.31",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "anthropic>=0.39",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
fedquery = "src.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
